refactor: decompose photos, program, weekly-summary, template screens (batch 3)
Extract hooks and sub-components from 4 large screen files:
- photos.tsx (702→132 lines): usePhotoActions, PhotoFilterHeader, PhotoMetaModal, PrivacyModal
- program/[id].tsx (671→285 lines): useProgramDetail, WeeklySchedule, ProgramHistory
- WeeklySummary.tsx (667→189 lines): useWeeklySummary, SummaryDetailSections
- template/[id].tsx (567→135 lines): useTemplateEditor, TemplateExerciseRow
