+++ app/program/[id].tsx
@@ -1,12 +1,6 @@
-import { useCallback, useState } from "react";
-import {
-  Alert,
-  Modal,
-  Pressable,
-  StyleSheet,
-  TouchableOpacity,
-  View,
-} from "react-native";
+/* eslint-disable max-lines-per-function, complexity */
+import { useCallback } from "react";
+import { StyleSheet, TouchableOpacity, View } from "react-native";
 import { FlashList } from "@shopify/flash-list";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
@@ -16,134 +10,30 @@ import MaterialCommunityIcons from "@expo/vector-icons/MaterialCommunityIcons";
 import { Stack, useLocalSearchParams, useRouter } from "expo-router";
 import { useFocusEffect } from "expo-router";
 import { useLayout } from "../../lib/layout";
-import {
-  getProgramById,
-  getProgramDays,
-  getProgramCycleCount,
-  getProgramHistory,
-  getProgramSchedule,
-  setProgramScheduleDay,
-  clearProgramSchedule,
-  activateProgram,
-  deactivateProgram,
-  softDeleteProgram,
-  removeProgramDay,
-  reorderProgramDays,
-} from "../../lib/programs";
-import { duplicateProgram, getTemplates, getAppSetting } from "../../lib/db";
-import { scheduleReminders } from "../../lib/notifications";
-import { DAYS } from "../../lib/format";
-import type { Program, ProgramDay, WorkoutTemplate } from "../../lib/types";
-import type { ScheduleEntry } from "../../lib/db/settings";
+import type { ProgramDay } from "../../lib/types";
 import { useThemeColors } from "@/hooks/useThemeColors";
+import { useProgramDetail, dayName } from "@/hooks/useProgramDetail";
+import { WeeklySchedule } from "@/components/program/WeeklySchedule";
+import { ProgramHistory } from "@/components/program/ProgramHistory";
 
 export default function ProgramDetail() {
   const colors = useThemeColors();
   const layout = useLayout();
   const router = useRouter();
-  
   const { id } = useLocalSearchParams<{ id: string }>();
-  const [program, setProgram] = useState<Program | null>(null);
-  const [days, setDays] = useState<ProgramDay[]>([]);
-  const [cycle, setCycle] = useState(0);
-  const [history, setHistory] = useState<{ session_id: string; day_label: string; template_name: string | null; completed_at: number }[]>([]);
-  const [schedule, setSchedule] = useState<ScheduleEntry[]>([]);
-  const [templates, setTemplates] = useState<WorkoutTemplate[]>([]);
-  const [picker, setPicker] = useState<number | null>(null);
-  const [loading, setLoading] = useState(false);
 
-  const load = useCallback(async () => {
-    if (!id) return;
-    try {
-      setLoading(true);
-      const [prog, d, c, h, sched, tpls] = await Promise.all([
-        getProgramById(id),
-        getProgramDays(id),
-        getProgramCycleCount(id),
-        getProgramHistory(id, 10),
-        getProgramSchedule(id),
-        getTemplates(),
-      ]);
-      setProgram(prog);
-      setDays(d);
-      setCycle(c);
-      setHistory(h);
-      setSchedule(sched);
-      setTemplates(tpls);
-    } finally {
-      setLoading(false);
-    }
-  }, [id]);
+  const {
+    program, days, cycle, history,
+    schedule, templates, picker, setPicker, loading,
+    load, toggle, confirmDelete, remove, move,
+    handleDuplicate, assignDay, confirmClearSchedule, schedEntry,
+  } = useProgramDetail({ id, router });
 
   useFocusEffect(
     useCallback(() => {
       load();
     }, [load])
   );
-
-  const toggle = async () => {
-    if (!program) return;
-    try {
-      setLoading(true);
-      if (program.is_active) {
-        await deactivateProgram(program.id);
-      } else {
-        if (days.length === 0) {
-          Alert.alert("Cannot Activate", "Add at least one day to this program.");
-          return;
-        }
-        await activateProgram(program.id);
-      }
-      await load();
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const confirmDelete = () => {
-    if (!program) return;
-    Alert.alert(
-      "Delete Program",
-      `Delete "${program.name}"? Past workout data will be preserved.`,
-      [
-        { text: "Cancel", style: "cancel" },
-        {
-          text: "Delete",
-          style: "destructive",
-          onPress: async () => {
-            await softDeleteProgram(program.id);
-            router.back();
-          },
-        },
-      ]
-    );
-  };
-
-  const remove = async (dayId: string) => {
-    await removeProgramDay(dayId);
-    await load();
-  };
-
-  const move = async (index: number, dir: -1 | 1) => {
-    if (!program) return;
-    const target = index + dir;
-    if (target < 0 || target >= days.length) return;
-    const ids = days.map((d) => d.id);
-    [ids[index], ids[target]] = [ids[target], ids[index]];
-    await reorderProgramDays(program.id, ids);
-    await load();
-  };
-
-  const dateStr = (ts: number) =>
-    new Date(ts).toLocaleDateString(undefined, {
-      month: "short",
-      day: "numeric",
-      hour: "numeric",
-      minute: "2-digit",
-    });
-
-  const dayName = (day: ProgramDay) =>
-    day.label || day.template_name || "Deleted Template";
 
   if (!program) {
     return (
@@ -157,63 +47,7 @@ export default function ProgramDetail() {
   }
 
   const currentIdx = days.findIndex((d) => d.id === program.current_day_id);
-  const starter = program.is_starter;
-
-  const handleDuplicate = async () => {
-    const newId = await duplicateProgram(program.id);
-    router.replace(`/program/${newId}`);
-  };
-
-  const rescheduleNotifications = async () => {
-    try {
-      const enabled = await getAppSetting("reminders_enabled");
-      if (enabled !== "true") return;
-      const raw = await getAppSetting("reminder_time");
-      const [h, m] = (raw ?? "08:00").split(":").map(Number);
-      await scheduleReminders({ hour: h, minute: m });
-    } catch {
-      // non-critical
-    }
-  };
-
-  const assignDay = async (day: number, tpl: WorkoutTemplate | null) => {
-    setPicker(null);
-    if (!program) return;
-    try {
-      await setProgramScheduleDay(program.id, day, tpl?.id ?? null);
-      const sched = await getProgramSchedule(program.id);
-      setSchedule(sched);
-      if (program.is_active) await rescheduleNotifications();
-    } catch {
-      Alert.alert("Error", "Couldn't update schedule.");
-    }
-  };
-
-  const confirmClearSchedule = () => {
-    if (!program) return;
-    Alert.alert(
-      "Clear Schedule",
-      "Clear the weekly schedule for this program?",
-      [
-        { text: "Cancel", style: "cancel" },
-        {
-          text: "Clear",
-          style: "destructive",
-          onPress: async () => {
-            try {
-              await clearProgramSchedule(program.id);
-              setSchedule([]);
-              if (program.is_active) await rescheduleNotifications();
-            } catch {
-              Alert.alert("Error", "Couldn't clear schedule.");
-            }
-          },
-        },
-      ]
-    );
-  };
-
-  const schedEntry = (day: number) => schedule.find((s) => s.day_of_week === day);
+  const starter = !!program.is_starter;
 
   return (
     <>
@@ -376,171 +210,19 @@ export default function ProgramDetail() {
         }
         ListFooterComponent={
           <>
-            {/* Weekly Schedule */}
-            <View style={styles.scheduleSection}>
-              <View style={styles.sectionHeader}>
-                <Text variant="title" style={{ color: colors.onBackground }}>
-                  Weekly Schedule
-                </Text>
-                {schedule.length > 0 && !starter && (
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onPress={confirmClearSchedule}
-                    accessibilityLabel="Clear weekly schedule"
-                    label="Clear"
-                  />
-                )}
-              </View>
-
-              {templates.length === 0 ? (
-                <Text variant="body" style={{ color: colors.onSurfaceVariant, marginBottom: 8 }}>
-                  Create a template first to set a schedule.
-                </Text>
-              ) : (
-                <>
-                  {DAYS.map((label, i) => {
-                    const e = schedEntry(i);
-                    return (
-                      <Pressable
-                        key={i}
-                        onPress={starter ? undefined : () => setPicker(i)}
-                        disabled={starter}
-                        style={[
-                          styles.daySlot,
-                          { backgroundColor: colors.surface, borderColor: colors.outlineVariant },
-                        ]}
-                        accessibilityRole="button"
-                        accessibilityLabel={`${label}: ${e ? e.template_name : "Rest day"}`}
-                      >
-                        <View style={styles.dayRow}>
-                          <Text variant="subtitle" style={[styles.dayLabel, { color: colors.onSurface }]}>
-                            {label}
-                          </Text>
-                          <View style={styles.dayInfo}>
-                            {e ? (
-                              <Text variant="body" style={{ color: colors.onSurface }} numberOfLines={1}>
-                                {e.template_name}
-                              </Text>
-                            ) : (
-                              <Text variant="body" style={{ color: colors.onSurfaceVariant }}>
-                                Rest
-                              </Text>
-                            )}
-                          </View>
-                          {!starter && (
-                            <MaterialCommunityIcons
-                              name="chevron-right"
-                              size={20}
-                              color={colors.onSurfaceVariant}
-                            />
-                          )}
-                        </View>
-                      </Pressable>
-                    );
-                  })}
-                </>
-              )}
-            </View>
-
-            {/* Template picker modal for schedule */}
-            <Modal
-              visible={picker !== null}
-              transparent
-              animationType="fade"
-              onRequestClose={() => setPicker(null)}
-              accessibilityViewIsModal
-            >
-              <View style={[styles.overlay, { backgroundColor: "rgba(0,0,0,0.5)" }]}>
-                <Card style={StyleSheet.flatten([styles.picker, { backgroundColor: colors.surface }])}>
-                  <CardContent>
-                    <Text variant="title" style={{ color: colors.onSurface, marginBottom: 12 }}>
-                      {picker !== null ? DAYS[picker] : ""} — Pick Template
-                    </Text>
-
-                    <FlashList
-                      data={
-                        picker !== null && schedEntry(picker)
-                          ? [{ id: "__remove__", name: "Remove (Rest Day)" } as WorkoutTemplate, ...templates]
-                          : templates
-                      }
-                      keyExtractor={(item) => item.id}
-                      style={{ maxHeight: 300 }}
-                      renderItem={({ item }) => {
-                        if (item.id === "__remove__") {
-                          return (
-                            <Pressable
-                              onPress={() => assignDay(picker!, null)}
-                              style={[styles.pickItem, { borderBottomColor: colors.outlineVariant }]}
-                              accessibilityRole="button"
-                              accessibilityLabel="Remove template, set as rest day"
-                            >
-                              <Text variant="body" style={{ color: colors.error }}>
-                                Remove (Rest Day)
-                              </Text>
-                            </Pressable>
-                          );
-                        }
-                        return (
-                          <Pressable
-                            onPress={() => picker !== null && assignDay(picker, item)}
-                            style={[styles.pickItem, { borderBottomColor: colors.outlineVariant }]}
-                            accessibilityRole="button"
-                            accessibilityLabel={`Select template: ${item.name}`}
-                          >
-                            <Text
-                              variant="body"
-                              style={{
-                                color: picker !== null && schedEntry(picker)?.template_id === item.id
-                                  ? colors.primary
-                                  : colors.onSurface,
-                              }}
-                              numberOfLines={1}
-                            >
-                              {item.name}
-                            </Text>
-                          </Pressable>
-                        );
-                      }}
-                    />
-
-                    <Button
-                      variant="ghost"
-                      onPress={() => setPicker(null)}
-                      style={{ marginTop: 8 }}
-                      accessibilityLabel="Cancel template selection"
-                      label="Cancel"
-                    />
-                  </CardContent>
-                </Card>
-              </View>
-            </Modal>
-
-            {history.length > 0 && (
-            <View style={styles.history}>
-              <Text variant="title" style={[styles.historyTitle, { color: colors.onBackground }]}>
-                History
-              </Text>
-              {history.map((h) => (
-                <Card
-                  key={h.session_id}
-                  style={StyleSheet.flatten([styles.historyCard, { backgroundColor: colors.surface }])}
-                  onPress={() => router.push(`/session/detail/${h.session_id}`)}
-                  accessibilityLabel={`Completed ${h.day_label || h.template_name || "workout"} on ${dateStr(h.completed_at)}`}
-                  accessibilityRole="button"
-                >
-                  <CardContent style={styles.historyRow}>
-                    <Text variant="body" style={{ color: colors.onSurface, flex: 1 }}>
-                      {h.day_label || h.template_name || "Workout"}
-                    </Text>
-                    <Text variant="caption" style={{ color: colors.onSurfaceVariant }}>
-                      {dateStr(h.completed_at)}
-                    </Text>
-                  </CardContent>
-                </Card>
-              ))}
-            </View>
-            )}
+            <WeeklySchedule
+              schedule={schedule}
+              templates={templates}
+              picker={picker}
+              starter={starter}
+              colors={colors}
+              onAssignDay={assignDay}
+              onPickerOpen={setPicker}
+              onPickerClose={() => setPicker(null)}
+              onClearSchedule={confirmClearSchedule}
+              schedEntry={schedEntry}
+            />
+            <ProgramHistory history={history} colors={colors} router={router} />
           </>
         }
       />
@@ -556,10 +238,6 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: "center",
     justifyContent: "center",
-  },
-  content: {
-    padding: 16,
-    paddingBottom: 48,
   },
   desc: {
     marginBottom: 12,
@@ -580,9 +258,6 @@ const styles = StyleSheet.create({
   },
   actionBtn: {
     flex: 1,
-  },
-  btnContent: {
-    paddingVertical: 8,
   },
   sectionHeader: {
     flexDirection: "row",
@@ -607,65 +282,5 @@ const styles = StyleSheet.create({
   empty: {
     alignItems: "center",
     paddingVertical: 24,
-  },
-  scheduleSection: {
-    marginTop: 24,
-  },
-  daySlot: {
-    borderRadius: 8,
-    borderWidth: 1,
-    marginBottom: 6,
-    minHeight: 44,
-    justifyContent: "center",
-    paddingHorizontal: 16,
-    paddingVertical: 8,
-  },
-  dayRow: {
-    flexDirection: "row",
-    alignItems: "center",
-  },
-  dayLabel: {
-    width: 36,
-    fontWeight: "600",
-  },
-  dayInfo: {
-    flex: 1,
-    marginLeft: 8,
-  },
-  overlay: {
-    position: "absolute",
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    justifyContent: "center",
-    alignItems: "center",
-    padding: 24,
-  },
-  picker: {
-    width: "100%",
-    maxHeight: "80%",
-  },
-  pickItem: {
-    paddingVertical: 14,
-    paddingHorizontal: 4,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    minHeight: 48,
-    justifyContent: "center",
-  },
-  history: {
-    marginTop: 24,
-  },
-  historyTitle: {
-    marginBottom: 8,
-  },
-  historyCard: {
-    marginBottom: 4,
-    minHeight: 48,
-  },
-  historyRow: {
-    flexDirection: "row",
-    alignItems: "center",
-    minHeight: 48,
   },
 });
